comps::PackageType: Move to separate header file
