+++ include/libdnf5/comps/group/package.hpp
@@ -20,61 +20,13 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #ifndef LIBDNF5_COMPS_GROUP_PACKAGE_HPP
 #define LIBDNF5_COMPS_GROUP_PACKAGE_HPP
 
-#include "libdnf5/common/exception.hpp"
+#include "package_type.hpp"
 
 #include <string>
-#include <vector>
-
 
 namespace libdnf5::comps {
 
-enum class PackageType : int {
-    CONDITIONAL = 1 << 0,  // a weak dependency
-    DEFAULT = 1 << 1,      // installed by default, but can be unchecked in the UI
-    MANDATORY = 1 << 2,    // installed
-    OPTIONAL = 1 << 3      // not installed by default, but can be checked in the UI
-};
-
-class InvalidPackageType : public libdnf5::Error {
-public:
-    InvalidPackageType(const std::string & type);
-    InvalidPackageType(const PackageType type);
-
-    const char * get_domain_name() const noexcept override { return "libdnf5::comps"; }
-    const char * get_name() const noexcept override { return "InvalidPackageType"; }
-};
-
-inline PackageType operator|(PackageType a, PackageType b) {
-    return static_cast<PackageType>(
-        static_cast<std::underlying_type<PackageType>::type>(a) |
-        static_cast<std::underlying_type<PackageType>::type>(b));
-}
-
-inline PackageType operator|=(PackageType & a, PackageType b) {
-    a = static_cast<PackageType>(
-        static_cast<std::underlying_type<PackageType>::type>(a) |
-        static_cast<std::underlying_type<PackageType>::type>(b));
-    return a;
-}
-
-inline constexpr PackageType operator&(PackageType a, PackageType b) {
-    return static_cast<PackageType>(
-        static_cast<std::underlying_type<PackageType>::type>(a) &
-        static_cast<std::underlying_type<PackageType>::type>(b));
-}
-
-inline constexpr bool any(PackageType flags) {
-    return static_cast<std::underlying_type<PackageType>::type>(flags) != 0;
-}
-
-PackageType package_type_from_string(const std::string & type);
-PackageType package_type_from_string(const std::vector<std::string> types);
-std::string package_type_to_string(const PackageType type);
-std::vector<std::string> package_types_to_strings(const PackageType types);
-
-
 // TODO(dmach): isn't it more a package dependency rather than a package?
-
 
 // @replaces dnf:dnf/comps.py:class:Package
 // @replaces dnf:dnf/comps.py:class:CompsTransPkg
@@ -122,6 +74,5 @@ private:
 
 
 }  // namespace libdnf5::comps
-
 
 #endif
